organized folders, cleaned up RegisterForm.js
