+++ src/components/Register.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import RegisterForm from './RegisterForm';
+export function Register() {
+  return (
+    <div><RegisterForm /></div>
+  )
+}
